refactor(cli): extract analyze parser args
